+++ blacklight_security/scanners/aws/__init__.py
@@ -1,5 +1,9 @@
 """AWS scanner implementations."""
 
+from .cloudtrail import CloudTrailScanner
+from .ec2 import EC2Scanner
+from .iam import IAMScanner
+from .rds import RDSScanner
 from .s3 import S3Scanner
 
-__all__ = ["S3Scanner"]
+__all__ = ["CloudTrailScanner", "EC2Scanner", "IAMScanner", "RDSScanner", "S3Scanner"]